New glossary plugin and related plugin js and css hooks
Glossary plugin allows addition of glossary items to the docs.
The plugin needed special javascript and css hooks, so I added the ability for the plugin
to request that special css and javascript items are inserted into the build so as to completely separate
the plugin from the main tool

Also some misc bug fixes and more special character handling support.