Added support for plugin javascript and css overrides
The glossary plugin needed the ability to set up javascript actions and styles.
Thus turnin adds that capability for a plugin to add unique configurations.

Also fixed a problem with SVG class overrides that was missed (and was resulting in incorrect file rendering).

Ran css files through a beautify script to make them easier to read.